Trying to hunt down propagation of mouse coordinates.
While still observing the jump, this does not show up anywhere in the
debug output.  While there are different "kinds" of coordinates used,
like (309.120, 13.263) in control.c, (307.120, 11.263) in darkroom.c,
(0.449, -0.208) and (596.729, -184) in masks.c for the same mouse_move
event, they do *not* show the jump of the shape.

[sk] .../control/control.c: dt_control_mouse_moved(309.120300, 13.263153, 1.000000, 0)
[sk] .../views/darkroom.c: mouse_moved(307.120300, 11.263153, 1.000000, 0)
[sk] .../develop/masks/masks.c:dt_masks_events_mouse_moved pzx = 0.449006, pzy = -0.208438
[sk] .../develop/masks/masks.c:dt_masks_events_mouse_moved posx = 596.729370, posy = -184.

I'm stuck when an event is called in `src/control/control.c` line 546.
But I assume that the central widget needs to be redrawn, tries to get
the shape's current location, and that's where it's messed up.